Rules refactor
