+++ Web/src/Rules/Dialogs/DeleteDialog.jsx
