+++ Web/src/Rules/RulesClasses/Rule.js
